requirements.txt
crewai==0.141.0
crewai-tools==0.51.1
ollama==0.1.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
validators==0.22.0
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-community==0.0.10
pydantic==2.5.0
plotly==6.2.0